Split Slave's main() into SlaveRunner
