+++ slave.bat
@@ -1,1 +1,1 @@
-mvnw -q compile exec:java -Dexec.mainClass="io.github.incplusplus.peerprocessing.slave.Slave" -Dexec.cleanupDaemonThreads=false
+mvnw -q compile exec:java -Dexec.mainClass="io.github.incplusplus.peerprocessing.slave.SlaveRunner" -Dexec.cleanupDaemonThreads=false
