+++ slave.sh
@@ -1,2 +1,2 @@
 #!/usr/bin/env bash
-./mvnw -q compile exec:java -Dexec.mainClass="io.github.incplusplus.peerprocessing.slave.Slave" -Dexec.cleanupDaemonThreads=false
+./mvnw -q compile exec:java -Dexec.mainClass="io.github.incplusplus.peerprocessing.slave.SlaveRunner" -Dexec.cleanupDaemonThreads=false
